Refactor motor control code and add documentation
This commit refactors the motor control code by moving the global variables and function declarations related to motor control to the "motor_control.h" file. It also adds documentation to explain the purpose of each variable and function.

The "main.cpp" file is the entry point of the project and contains the initialization and main loop logic.

The "motor_control.cpp" file implements the functions declared in "motor_control.h".

Development was done using PlatformIO and the Arduino framework.

Installation and usage instructions are provided in the readme file.